feat: apple 모듈 auth 모듈로 병합
