+++ src/auth/auth.controller.ts
@@ -6,6 +6,7 @@ import {
   Req,
   Put,
   Body,
+  Post,
 } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthGuard } from '@nestjs/passport';
@@ -20,6 +21,7 @@ import {
   ApiBearerAuth,
 } from '@nestjs/swagger';
 import { FirstLoginDto } from './dtos/firstLogin-dto';
+import { AppleLoginDto } from './dtos/apple-login-dto';
 
 @Controller('auth')
 @ApiTags('인증 인가')
@@ -66,5 +68,27 @@ export class AuthController {
       req.user.id,
       req.headers.authorization.substring(7),
     );
+  }
+
+  // -------------------------- 애플 --------------------------------
+  //애플 로그인
+  @Post('/apple')
+  @ApiOperation({
+    summary: '애플 sign in',
+  })
+  @ApiResponse({ status: 201, description: '애플 로그인 성공' })
+  async login(@Body() appleLoginDto: AppleLoginDto) {
+    return this.authService.appleLogin(appleLoginDto.authId);
+  }
+  //애플에서 유저가 "이메일 변경, 앱 서비스 해지, 애플 계정 탈퇴"를 했을 경우,
+  //App ID apple sign in 에서 입력한 Endpoint URL로 유저 정보와 이벤트에 대한 PAYLOAD 데이터를 전송.
+  @Post('/apple-endpoint')
+  @ApiOperation({
+    summary: '애플 유저 관련 공지 endpoint',
+    description:
+      '애플에서 유저가 "이메일 변경, 앱 서비스 해지, 애플 계정 탈퇴"를 했을 경우',
+  })
+  async endpoint() {
+    //추후 논의 후 구현.
   }
 }
